requirements.txt
certifi==2019.6.16
chardet==3.0.4
cloudinary==1.16.0
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.17
django-cors-headers==3.0.2
django-material==1.1.1
django-render-block==0.6
django-rest-swagger==2.2.0
django-richtextfield==1.4.0
django-templated-email==2.3.0
djangorestframework==3.10.1
djangorestframework-jwt==1.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mock==3.0.5
openapi-codec==1.3.2
psycopg2-binary==2.8.3
PyJWT==1.7.1
pyotp==2.3.0
pytz==2019.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.3
